Refactor install scripts to remove redudancy
