+++ cmdstanpy/install_cmdstan.py
@@ -16,7 +16,6 @@ Optional command line arguments:
    -c, --compiler : flag, add C++ compiler to path (Windows only)
 """
 import argparse
-import contextlib
 import json
 import os
 import platform
@@ -29,10 +28,10 @@ import urllib.request
 from collections import OrderedDict
 from pathlib import Path
 from time import sleep
-from typing import Callable, Dict, Iterator, Optional
+from typing import Callable, Dict, Optional
 
 from cmdstanpy import _DOT_CMDSTAN, _DOT_CMDSTANPY
-from cmdstanpy.utils import validate_dir
+from cmdstanpy.utils import pushd, validate_dir, wrap_progress_hook
 
 EXTENSION = '.exe' if platform.system() == 'Windows' else ''
 
@@ -43,15 +42,6 @@ class CmdStanRetrieveError(RuntimeError):
 
 class CmdStanInstallError(RuntimeError):
     pass
-
-
-@contextlib.contextmanager
-def pushd(new_dir: str) -> Iterator[None]:
-    """Acts like pushd/popd."""
-    previous_dir = os.getcwd()
-    os.chdir(new_dir)
-    yield
-    os.chdir(previous_dir)
 
 
 def usage() -> None:
@@ -238,34 +228,6 @@ def latest_version() -> str:
     if match is not None:
         tag = match.group(1)
     return tag  # type: ignore
-
-
-def wrap_progress_hook() -> Optional[Callable[[int, int, int], None]]:
-    try:
-        from tqdm import tqdm
-
-        pbar = tqdm(
-            unit='B',
-            unit_scale=True,
-            unit_divisor=1024,
-        )
-
-        def download_progress_hook(
-            count: int, block_size: int, total_size: int
-        ) -> None:
-            if pbar.total is None:
-                pbar.total = total_size
-                pbar.reset()
-            downloaded_size = count * block_size
-            pbar.update(downloaded_size - pbar.n)
-            if pbar.n >= total_size:
-                pbar.close()
-
-    except (ImportError, ModuleNotFoundError):
-        print("tqdm is not installed, progressbar not shown")
-        return None
-
-    return download_progress_hook
 
 
 def retrieve_version(version: str, progress: bool = True) -> None:
